Add support for deploying teleport on CoreOS
The teleport binary for Linux runs on CoreOS and systemd is available
for managing the service lifecycle. We extend the plugin to support
CoreOS by adding it to the list of supported operating system in
config.pp.

We expose other confguration parameters (archive_url, systemd_file) to
make the installation more flexible.

The archive module has been added as a direct dependency since it's
being used by the module.